fix column display of data
For some reason in qml what used to be columns is mapped to roles now.
Normally you would expect the QModelIndex to carry the column number but
it always seems to be 0. So we use the shortcut to index by roles, which
is messy but is a quick fix for this example.